+++ src-dbg/flow_storm/debugger/events_processor.clj
@@ -10,8 +10,7 @@
             [flow-storm.debugger.ui.flows.general :as ui-general]
             [flow-storm.debugger.ui.utils :as ui-utils]
             [flow-storm.debugger.ui.data-windows.data-windows :as data-windows]
-            [flow-storm.debugger.state :as dbg-state]
-            [flow-storm.utils :refer [log]]))
+            [flow-storm.debugger.state :as dbg-state]))
 
 (defn- vanilla-var-instrumented-event [{:keys [var-ns var-name]}]
   (ui-utils/run-later
@@ -102,14 +101,10 @@
 (defn- multi-timeline-recording-updated-event [{:keys [recording?]}]
   (flows-screen/set-multi-timeline-recording-btn recording?))
 
-(defn- function-unwinded-event [unwind-data]
-  (let [ui-unwinds-limit 200]
-    (if (< (count (dbg-state/get-fn-unwinds)) ui-unwinds-limit)
-      (do
-        (dbg-state/add-fn-unwind unwind-data)
-        (ui-utils/run-later
-         (flows-screen/update-exceptions-combo)))
-      (log (format "Functions unwinds limit of %d exceeded, not adding more exceptions to the Exceptions menu." ui-unwinds-limit)))))
+(defn- function-unwinded-event [{:keys [flow-id] :as unwind-data}]
+  (dbg-state/add-fn-unwind unwind-data)
+  (ui-utils/run-later
+    (flows-screen/update-exceptions-combo flow-id)))
 
 (defn data-window-push-val-data-event [{:keys [dw-id val-data root?]}]
   (data-windows/push-val dw-id val-data root?))